Fix __PHP_Incomplete_Class when not logged in
When no user is logged in and I want to run any query what so ever I receive a fatal error of $userData being and __PHP_Incomplete_Class object.

When an user is logged in I no longer have this problem. The proposed code fixes this when the object is broken only.