Screen for whether the 'objects' field is set under add or remove relations
Trying to remove relations throws an error in the current ParseServer, preventing a removal. Both the add & remove relation arrays always pass the 'empty' check due to the following lines in __construct:
```
$this->relationsToAdd['null'] = [];
$this->relationsToRemove['null'] = [];
```
On trying to add a relation nothing bad seems happens, as the first operation succeeds. When trying to remove a relation however the action never occurs as it will fail and stop executing on the first element, which does not contain an 'objects' array. Note that this does not happen if both removing and adding objects to a relation. Parse Server currently has no check for this so it just dies with a TypeError of 'Cannot read property 'className' of undefined'.

This can be observed by flipping the order of elements here, causing whichever one is first to either succeed or trigger a failure if not set :
```
return [
    '__op' => 'Batch',
    'ops'  => [$addRelation, $removeRelation],
];
```